added design editor to new site system
